Modernize the state of the Git repository
Use an automatic Oomph setup to set up the development environment. It's
currently authored here:

https://git.eclipse.org/c/oomph/org.eclipse.oomph.git/tree/setups/interim/CBI.setup

Consistently use EPL 2.0 everywhere.

Use consistent copyright headers with consistent licenses in all files
that should have them, e.g., all *.java files.

Avoid formatting of Javadoc in projects with EMF models.  This
formatting corrupts the XML comment markers such that the EMF generator
can no longer merge the Javadoc.  Fix all the corrupted Javadoc
instances such that regeneration works properly again.

Move all the projects to BREE Java 11.

Regenerate all the models properly, including generating the plugin.xml
content properly with EMF markers that support merging in the future.
Fix all the problems in the models that cause errors in the regenerated
code.

Use MinimalEObjectImpl$Container for all the generated models.

Update and simplify the target platform to use the latest (2021-12)
dependencies.  It is generated from the Oomph setup.

Update the versions of all the plugins, features, and poms (as well as
the product) to 1.0.400.

Put the editor in a properly named category, not the EMF examples
category.

Consistently name all the plugins, features, and categories for a more
consistent presentation to the end user.

Remove pack200 from the Tycho build and eliminate warnings for source
feature generation.

Filter the root projects contents to avoid duplicate resources in the
workspace.

Filter target folders so that local Tycho build results are not visible
to search and avoid indexing overhead.

The experimental Nexus indexer doen't compile against the jars in the
most recent m2e and even when the older ones it was using, did not
actually function.  I've commented out parts of it for now such that
only the slower crawling index works.